requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
gunicorn>=21.2.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic[email]
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
geoalchemy2>=0.14.0

# AWS SDK
boto3>=1.26.0

# Google Cloud SDK
google-cloud-storage>=2.10.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utils
python-dotenv>=1.0.0
pyopenssl>=23.0.0
python-slugify>=8.0.0
orjson>=3.9.10
ujson>=5.8.0
aiofiles>=23.2.1
httpx>=0.26.0